Refactor to avoid exporting hcl2 entities
