+++ pkg/codegen/hcl2/binder_schema_test.go
@@ -1,11 +1,10 @@
-package hcl2_test
+package hcl2
 
 import (
 	"path/filepath"
 	"testing"
 
-	"github.com/pulumi/pulumi/pkg/v3/codegen/hcl2"
-	"github.com/pulumi/pulumi/pkg/v3/codegen/internal/test"
+	"github.com/pulumi/pulumi/pkg/v3/codegen/internal/utils"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/schema"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 )
@@ -13,10 +12,10 @@ import (
 var testdataPath = filepath.Join("..", "internal", "test", "testdata")
 
 func BenchmarkLoadPackage(b *testing.B) {
-	loader := schema.NewPluginLoader(test.NewHost(testdataPath))
+	loader := schema.NewPluginLoader(utils.NewHost(testdataPath))
 
 	for n := 0; n < b.N; n++ {
-		_, err := hcl2.NewPackageCache().LoadPackageSchema(loader, "aws")
+		_, err := NewPackageCache().loadPackageSchema(loader, "aws")
 		contract.AssertNoError(err)
 	}
 }
